wayland: Use pop shell type for all transient style of surfaces
I've tested the behavior of all sub-windows in Chrome and Content Shell on X11
and all seemed to grab the input pointer (in a popup style) when opened.
Therefore this commit changes our implementation to work accordingly.

It's easy to test this functionality by opening a sub-window (e.g. right click
on the Web content), and without closing it go outside the Chromium window and
hover the mouse cursor over another Wayland application. Now, none interaction
using the mouse pointer should happen on those applications unless you click
again.

In other words, input grabbing across applications now are alright because
Wayland does it implicitly but there still some in-application logic to be
done. This is coming in the following commit.